# qa-shop/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: QA Shop Practice API
  version: "1.0.0"
  description: |
    Test otomasyonu pratiği için gerçek bir e-ticaret API'si. Mock değil:
    veri kalıcı, işlemler transaction içinde, stok gerçekten düşüyor.

    **Sandbox anahtarı göndermezsen** demo verisine salt okunur bağlanırsın —
    `GET` uçları çalışır, yazma denemeleri `401` döner. İlk denemeyi
    kolaylaştırmak için böyle.

    **Kendi alanını açmak için:** `POST /sandbox` → dönen `apiKey` değerini
    `X-Sandbox-Key` başlığında gönder. Alan 7 gün sonra otomatik silinir.

    **Her test koşumundan önce** `POST /sandbox/reset` çağır — temiz durum
    otomasyonun en temel disiplinidir ve burada bir uç olarak hazır.

    Demo hesabı: `demo@qashop.test` / `Password123!`
  contact:
    name: LearnQA.dev
    url: https://learnqa.dev
servers:
  - url: http://localhost:4000
    description: Lokal Docker (docker compose up)
...
security:
  - SandboxKey: []
paths:
...
  /api/v1/carts:
    post:
      tags: [Sepet]
      summary: Sepet aç
      description: Token gönderirsen sepet kullanıcıya bağlanır; göndermezsen misafir sepeti olur.
      responses:
        "201":
          description: Sepet oluşturuldu
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart: { $ref: "#/components/schemas/Cart" }
        "401": { $ref: "#/components/responses/ReadOnlySandbox" }
  /api/v1/carts/{id}:
    get:
      tags: [Sepet]
      summary: Sepet detayı ve toplamlar
      description: |
        `warnings.priceChanged` alanı, sepetteki fiyatın ürünün güncel
        fiyatından farklı olduğu satırları listeler. Sepet eski fiyatı korur —
        bu bir bug değil, bilinçli bir karardır; ama arayüzün hangisini
        göstereceğine kendi karar verdiği için TEST EDİLMESİ gereken bir karardır.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Sepet
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CartDetail" }
        "403": { description: Sepet başka bir kullanıcıya ait }
        "404": { description: Sepet bulunamadı }
  /api/v1/carts/{id}/items:
    post:
      tags: [Sepet]
      summary: Sepete ürün ekle
      description: |
        Stok REZERVE edilir (`reserved_qty` artar). Rezervasyon olmadan iki
        kullanıcı son ürünü sepetine atar ve hata ancak ödeme sonrası çıkar.

        Aynı varyant ikinci kez eklenirse adet artırılır, yeni satır açılmaz.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variantId]
              properties:
                variantId: { type: integer, example: 1 }
                qty: { type: integer, minimum: 1, default: 1, example: 2 }
      responses:
        "201":
          description: Eklendi, güncel sepet döner
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CartDetail" }
        "409":
          description: |
            `OUT_OF_STOCK` — `details.available` kaç adet kaldığını söyler.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`INVALID_VARIANT`, `INVALID_QTY` veya `PRODUCT_INACTIVE`"
        "404": { description: Varyant bulunamadı }
  /api/v1/carts/{id}/items/{itemId}:
    patch:
      tags: [Sepet]
      summary: Satır adedini güncelle
      description: |
        `qty: 0` ile silme YAPILMAZ → `422`. Silmek için `DELETE` var. Aynı işi
        iki farklı yolla yapmak, iki davranışın zamanla ayrışmasına yol açar.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: itemId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qty]
              properties:
                qty: { type: integer, minimum: 1, example: 3 }
      responses:
        "200": { description: Güncellendi, güncel sepet döner }
        "409": { description: "`OUT_OF_STOCK`" }
        "422": { description: "`INVALID_QTY`" }
        "404": { description: Sepet satırı bulunamadı }
    delete:
      tags: [Sepet]
      summary: Satırı sepetten çıkar
      description: Rezervasyon serbest bırakılır. Bu adım atlanırsa stok sızar.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: itemId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Silindi }
        "404": { description: Sepet satırı bulunamadı }
  /api/v1/carts/{id}/coupon:
    post:
      tags: [Sepet]
      summary: Kupon uygula
      description: |
        Reddetme nedeni AYRI bir kodla döner — "kupon geçersiz" demek hangi
        kuralın devreye girdiğini test edilemez kılardı:

        | Kod | Anlamı | Tohum veride örnek |
        |---|---|---|
        | `COUPON_NOT_FOUND` | Böyle bir kupon yok | — |
        | `COUPON_NOT_STARTED` | Henüz başlamadı | `FUTURE15` |
        | `COUPON_EXPIRED` | Süresi doldu | `EXPIRED20` |
        | `COUPON_USAGE_LIMIT_REACHED` | Limit doldu | `MAXEDOUT` |
        | `COUPON_MIN_TOTAL_NOT_MET` | Alt tutar yetersiz | `VIP1000` |

        Geçerli örnek: `WELCOME10` (%10), `SAVE50` (50 TL, min 300).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, example: WELCOME10 }
      responses:
        "200": { description: Kupon uygulandı, güncel sepet döner }
        "422":
          description: Kupon reddedildi (kod yukarıdaki tablodan)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ─────────────────────────────────── Sipariş ───────────────────────────────
...
components:
  securitySchemes:
    SandboxKey:
      type: apiKey
      in: header
      name: X-Sandbox-Key
      description: |
        Hangi veri alanına bağlanacağını belirler. `POST /api/v1/sandbox` ile
        alınır. Gönderilmezse demo verisine SALT OKUNUR bağlanılır.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /api/v1/auth/login` ile alınır. Gerçek JWT formatındadır —
        jwt.io'ya yapıştırıp içindeki `sub`, `sandbox`, `jti`, `exp`
        alanlarını görebilirsin.
  responses:
...
    ReadOnlySandbox:
      description: Demo verisi salt okunurdur — önce `POST /api/v1/sandbox`
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: |
        Hata gövdesi HER uçta aynı şekildedir. Her endpoint kendi hata şeklini
        döndürseydi test kodu her uç için ayrı yazılır ve kırılgan olurdu.
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string, example: Stokta yalnızca 2 adet var }
            details:
              type: object
              nullable: true
              example: { requested: 5, available: 2 }
        correlationId:
          type: string
          example: req-8f3c1e2a-...
          description: Cevabın `X-Correlation-Id` başlığında da döner.
...
    Cart:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer, nullable: true }
        guest_token: { type: string, nullable: true }
        status: { type: string, enum: [open, converted, abandoned] }
        created_at: { type: string, format: date-time }
    CartItem:
      type: object
      properties:
        id: { type: integer }
        variant_id: { type: integer }
        variant_sku: { type: string }
        product_id: { type: integer }
        product_name: { type: string }
        size: { type: string }
        color: { type: string }
        qty: { type: integer }
        unit_price:
          type: number
          description: Sepete atıldığı ANDAKİ fiyat (snapshot).
        current_price:
          type: number
          description: Ürünün ŞU ANKİ fiyatı. unit_price ile farklıysa uyarı üretilir.
        line_total: { type: number }
        stock_qty: { type: integer }
        available: { type: integer }
    Totals:
      type: object
      description: |
        grand_total = subtotal - discount_total + shipping_total
        Kargo 500 TL ve üzeri bedava, altında 29.90.
      properties:
        subtotal: { type: number }
        discount_total: { type: number }
        shipping_total: { type: number }
        grand_total: { type: number }
    CartDetail:
      type: object
      properties:
        cart: { $ref: "#/components/schemas/Cart" }
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        itemCount: { type: integer }
        totals: { $ref: "#/components/schemas/Totals" }
        warnings:
          type: object
          properties:
            priceChanged:
              type: array
              items:
                type: object
                properties:
                  itemId: { type: integer }
                  variantSku: { type: string }
                  cartPrice: { type: number }
                  currentPrice: { type: number }
...